Separate simulation slider and unify UI with BaseSelectionSlider
- Created BaseSelectionSlider.tsx in parts to encapsulate shared slider-input logic
- Refactored OptionSliderControl to use BaseSelectionSlider
- Created SimulationSliderControl specifically for the simulation feature
- Updated SimulationDialog to use SimulationSliderControl
- Deleted obsolete useSelectionSlider hook as logic is now in the base component
- Ensured strict separation between option-specific blocks and simulation features while maintaining UI consistency
- Fixed Biome linting and formatting issues
